Rotating around something

+++ OpenGL/src/maths/Maths.cpp
@@ -3,11 +3,11 @@
 namespace Maths {
 
 	float radians(float degrees) {
-		return degrees * PI_OVER_ONE_EIGHTY;
+		return degrees * static_cast<float>(PI_OVER_ONE_EIGHTY);
 	}
 
 	float degrees(float radians) {
-		return radians * ONE_EIGHTY_OVER_PI;
+		return radians * static_cast<float>(ONE_EIGHTY_OVER_PI);
 	}
 
 }
